requirements.txt
Django>=1.6.1
django-allauth>=0.16.1
twython>=3.2.0
facebook-sdk==1.0.0-alpha.1 #https://github.com/ejbdev/facebook-sdk/archive/1.0.0-alpha.1.tar.gz
